# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cotopia Streaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Songs ─────────────────────────────────────────────────────────────────
...
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # ── Favorites ─────────────────────────────────────────────────────────────
...
  /favorites/songs/{id}:
    post:
      operationId: favoriteSong
      tags: [favorites]
      summary: Add a song to favorites
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
    delete:
      operationId: unfavoriteSong
      tags: [favorites]
      summary: Remove a song from favorites
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
...
  /favorites/videos/{id}:
    post:
      operationId: favoriteVideo
      tags: [favorites]
      summary: Add a video to favorites
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
    delete:
      operationId: unfavoriteVideo
      tags: [favorites]
      summary: Remove a video from favorites
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK

  # ── Submissions ───────────────────────────────────────────────────────────
...
  /company/posts:
    get:
      operationId: listCompanyPosts
      tags: [company]
      summary: List company posts/announcements
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [announcement, video, product_update, artist_spotlight, label_spotlight, campaign]
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyPost"
    post:
      operationId: createCompanyPost
      tags: [company]
      summary: Create a company post (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyPost"
  /company/posts/{id}:
    get:
      operationId: getCompanyPost
      tags: [company]
      summary: Get a company post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyPost"
    patch:
      operationId: updateCompanyPost
      tags: [company]
      summary: Update a company post (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyPostUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyPost"
    delete:
      operationId: deleteCompanyPost
      tags: [company]
      summary: Delete a company post (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # ── Admin ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, username, role]
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [listener, artist, label, business]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, username, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [listener, artist, label, business, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string

    # ── Songs ───────────────────────────────────────────────────────────────
...
    CompanyPost:
      type: object
      required: [id, title, type, content, isPinned, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [announcement, video, product_update, artist_spotlight, label_spotlight, campaign]
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        isPinned:
          type: boolean
        createdAt:
          type: string
    CompanyPostInput:
      type: object
      required: [title, type, content]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [announcement, video, product_update, artist_spotlight, label_spotlight, campaign]
        content:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        isPinned:
          type: boolean
    CompanyPostUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        isPinned:
          type: boolean

    # ── Admin ─────────────────────────────────────────────────────────────────
...