# lib/api-zod/src/generated/types/companyPostUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cotopia Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyPostUpdate {
  title?: string;
  content?: string;
  imageUrl?: string;
  videoUrl?: string;
  isPinned?: boolean;
}